fix(test): comprehensive tier-2 sweep — move all thread/async/timing tests out of unit/
Systematic scan of tests/unit/ found 11 test files (or partial files) that use
AsyncLocalRuntime thread pools, threading.Thread, asyncio.run(), timing assertions,
or concurrent I/O patterns incompatible with CI unit-test constraints.

Full file moves (7):
- test_verifier_thread_safety.py — 20-thread stress tests
- test_audit_thread_safety.py — ThreadPoolExecutor concurrency
- test_verifier_security_hardening.py — asyncio.run() throughout
- test_async_runtime_production.py — AsyncLocalRuntime + timing
- test_connection_manager_events.py — 100% ConnectionManager I/O
- test_connection.py (db/) — 100% ConnectionManager I/O
- test_core_runtime_injection.py — AsyncLocalRuntime in every test

Partial file splits (4, extract tier-2 classes, keep unit remainder):
- test_runtime_lifecycle.py — async half extracted
- test_template_parameter_resolution.py — async runtime tests extracted
- test_sqlite_invariants.py — WAL/pool/stress tests extracted
- test_infrastructure_validation.py — PerformanceTracker timing extracted

Unit suite: 5412 passed, 0 failures, 20s. Tier-2 suite: 451 passed, 0 failures.
