+++ tests/tier2_integration/runtime/test_connection_manager_events.py
